requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
pycparser==2.20
PyJWT==1.7.1
pyotp==2.4.1
pyparsing==2.4.6
PyQRCode==1.2.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
simplejson==3.17.2
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
twilio==6.50.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
